fix(xiqse_workflow): use Long for executionId to match schema
XIQ-SE GraphQL schema types workflows.execution(executionId) as Long,
not Int. Using Int! caused a VariableTypeMismatch validation error.

+++ plugins/module_utils/xiqse.py
@@ -304,7 +304,7 @@ class XIQSE:
           @staticmethod
           def execution():
             return """
-              query Workflow($executionId: Int!) {
+              query Workflow($executionId: Long!) {
                 workflows {
                   execution(executionId: $executionId) {
                     status
